Added management of firewalld to init
Supporting enabled so that you can use foreman and apply this to a
parent hostgroup while leaving it disabled, then make changes on a child
prior to enabling the actual package.  More of a foreman process quirk.

Other option is to add the module to the child group and do its setup
and then do parent.  But that could cause issues if you ever need to
temporarily remove it from a child for some reason.(Foreman inheritance
FTW!)