Added the ability to submit our hashes inside defined types as well.
This allows us to still use our custom function for making firewalld easier to use, while also allowing the profiles part of roles/profiles added as a class in foreman to co-exist with using the class itself in foreman which allows overrides.
This will likely only be used during testing out changes until we can get a stable deployment platform.